BRK and RTI instructions
the BRK instruction is still "cheating": it does NOT set the PC yet to the 0xFFFE interrupt vector
address!
